add AbstractStoreFileManager and Refactor the setter logic in PersistedStoreFileManager
